index Servers by SocketAddr
This commit adds support for watching Server resources and storing an
index mapping SocketAddrs to Server watches. The SocketAddrs that refer
to a Server are determined using the existing index of pods.

Additionally, I've added support for dumping the entire start of the
index to stdout periodically. This is useful since we don't currently
have an API to serve, and there ought to be some way to see the
prototype controller's current view of the cluster's state.

A subsequent commit will also index HTTPRoutes and associate them with
the indexed Servers.