add `Index::addr_server_rx` (#4)
This commit adds an `Index::addr_server_rx`, similar to the
policy-controller's `Index::pod_server_rx`. This will be the main entry
point for looking up the index that's used by the gRPC API.

@adleong, I went ahead and added this in a separate, small branch, so
that once it merges, we can (hopefully) work on implementing the gRPC
API and additional indexing changes in parallel without touching the
same files at all.