index HTTPRoutes and bind them to Servers (#3)
This branch adds support for indexing HTTPRoute resources to the
prototype client-policy controller. Next, we'll want to also index the
new CRDs and bind them to HTTPRoutes as well, but that will happen in my
next PR.

I also fixed a bug where the controller didn't populate pods with
default servers when they were created. I should have done this in a
separate branch, but 🤷‍♀️

![image](https://user-images.githubusercontent.com/2796466/189503986-443a3a90-33fc-4e25-8274-0f52ac23610b.png)