implement ClientPolicyBinding (#10)
This branch rewrites the index in order to implement
ClientPolicyBinding. Now, rather than indexing pods as the "primary"
indexed object, we make Services the primary object in the index. An
`OutboundService` now tracks the ClientPolicyBindings for its
HTTPClientPolicies, and the routes on that service track their bindings.
This way, when we have a watch on an `OutboundService`, we can determine
which policies on that service apply to a particular client based on the
binding's pod selector.

Policies are now only bound to service ClusterIPs.

Pod indexing is temporarily removed. The next step will be to put back
pod indexing so that we can determine which policies apply to a client's
workload identifier. Each connection to the client policy controller's
discovery endpoint will hold a watch on the `OutboundService` for the
requested address, and the pod metadata for the client. Then, whenever
the `OutboundService` watch is updated, we'll determine which policies
apply to that client and send them to the proxy that opened the lookup.