add pod index for looking up pod meta, gRPC plumbing
This branch is the first step towards actually serving a gRPC discovery
API. It re-adds a simple pod index which just indexes a pod's labels by
pod ns/name, so that we can determine which policies bind to a client
pod.

I've added a stub gRPC server that, when receiving a
`GetDestination` request, looks up the pod that issued the request and
the `OutboundService` for that address, and builds an `async_stream`
with watches for the pod and service. This way, we can start actually
sending discovery updates when the pod's labels or the service whose IP
it requested change.

I haven't actually written the code for constructing the gRPC messages
yet, so right now, the endpoint always returns "not yet implemented" if
the requested address/client context exists. The next step will be to
actually make a gRPC message.